backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==4.0.1